requirements.txt
streamlit
qrcode
pillow
pyzbar
pandas
